feat: add read all by id functionality for tags, collections, and links
allow tags of a user to be read, all collections to be read, and all links to be created correctly

+++ src/services/collection/index.js
@@ -1,6 +1,6 @@
 import { create, validate } from './create';
-import { readById, readAll } from './read';
+import { readById, readAllByUserId, readAll } from './read';
 import { deleteById } from './delete';
 import { update } from './update';
 
-export { create, validate, readById, deleteById, update, readAll };
+export { create, validate, readById, readAllByUserId, deleteById, update, readAll };
